+++ ft_memcmp.c
@@ -6,11 +6,13 @@
 /*   By: ytoro-mo <ytoro-mo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/19 17:37:50 by ytoro-mo          #+#    #+#             */
-/*   Updated: 2022/04/19 17:59:48 by ytoro-mo         ###   ########.fr       */
+/*   Updated: 2022/04/20 12:20:31 by ytoro-mo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_memcmp(const void *s1, const void *s2, unsigned long n)
+#include "libft.h"
+
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
 	while (*(unsigned char *)s1++ != '\0' && *(unsigned char *)s2++ != '\0'
 		&& n--)
