+++ ft_strncmp.c
@@ -6,13 +6,15 @@
 /*   By: ytoro-mo <ytoro-mo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/25 12:54:32 by ytoro-mo          #+#    #+#             */
-/*   Updated: 2022/04/19 17:16:39 by ytoro-mo         ###   ########.fr       */
+/*   Updated: 2022/04/20 12:19:36 by ytoro-mo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strncmp(char *s1, char *s2, unsigned long n)
+#include "libft.h"
+
+int	ft_strncmp(char *s1, char *s2, size_t n)
 {
-	unsigned long	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n && (s1[i] != '\0' && s2[i] != '\0'))
