+++ ft_strnstr.c
@@ -6,11 +6,13 @@
 /*   By: ytoro-mo <ytoro-mo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/20 10:08:34 by ytoro-mo          #+#    #+#             */
-/*   Updated: 2022/04/20 10:44:19 by ytoro-mo         ###   ########.fr       */
+/*   Updated: 2022/04/20 12:20:56 by ytoro-mo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-char	*ft_strnstr(const char *haystack, const char *needle, unsigned long len)
+#include "libft.h"
+
+char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
 	char	*ptr;
 	int		i;
